alterar quantidade no carrinho

+++ projetoFinalThamiresBeatriz/src/stores/carrinho.js
@@ -9,7 +9,9 @@ export const useCarrinhoStore = defineStore({
   actions: {
     async getAllCarrinho() {
       try {
-        const { data } = await axios.get("http://localhost:4001/carrinho?expand=livros");
+        const { data } = await axios.get(
+          "http://localhost:4001/carrinho?expand=livros"
+        );
         this.carrinho = data;
         return Promise.resolve();
       } catch (e) {
@@ -19,24 +21,46 @@ export const useCarrinhoStore = defineStore({
         return Promise.reject("Erro desconhecido ao consultar 'Livros'");
       }
     },
- 
+
     async addCarrinho(livro_id) {
-        try {
-          const { data } = await axios.post( "http://localhost:4001/carrinho", livro_id);
+      try {
+        const idlivro = livro_id.livrosId;
+        const index = this.carrinho.findIndex(
+          (livros) => livros.livrosId === idlivro
+        );
+        //console.log(index);
+
+        if (index >= 0) {
+         // console.log("existe");
+          const dado = { quantidade: this.carrinho[index].quantidade + 1 };
+          await axios.patch(`http://localhost:4001/carrinho/${index}`, dado);
+          return Promise.resolve("Adicionado ao carrinho!");
+        } else {
+          const { data } = await axios.post("http://localhost:4001/carrinho",livro_id);
           this.carrinho.push(data);
           return Promise.resolve("Adicionado ao carrinho!");
-        } catch (e) {
-          console.error(e);
-          return Promise.reject(e);
         }
-      },       
+      } catch (e) {
+        console.error(e);
+        return Promise.reject(e);
+      }
+    },
     async deleteCarrinho(livros_id) {
       try {
-        await axios.delete(`http://localhost:4001/carrinho/${livros_id}`);
+        //   await axios.delete(`http://localhost:4001/carrinho/${livros_id}`);
         const index = this.carrinho.findIndex(
           (livros) => livros.id === livros_id
         );
-        this.carrinho.splice(index, 1);
+        if (this.carrinho[index].quantidade > 1) {
+          const dado = { quantidade: this.carrinho[index].quantidade - 1 };
+          await axios.patch(
+            `http://localhost:4001/carrinho/${livros_id}`,
+            dado
+          );
+        } else {
+          await axios.delete(`http://localhost:4001/carrinho/${livros_id}`);
+        }
+        //  this.carrinho.splice(index, 1);
         return Promise.resolve();
       } catch (e) {
         return Promise.reject("Erro ao excluir");
@@ -44,4 +68,3 @@ export const useCarrinhoStore = defineStore({
     },
   },
 });
-
